+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * [#101](https://github.com/sdss/clu/issues/101) The legacy actor now accepts command strings with a commander id (e.g., `APO.Jose 10 status`). The commander id is stored in the `Command` object. Added a `Command.send_command()` that will call the remote command propagating the commander ID. For example, if a command has commander `APO.Jose` and sends a command to `guider`, the `guider` actor will receive a command with commander `APO.Jose.guider`.
 
+* If not specified, a message with code `e` or `f`, or an exception, will use the keyword `error`.
+
 
 ## 1.5.0 - October 12, 2021